Move data container files into same directory.
